FudgeC2/requirements.txt
flask==1.0.2
flask_sqlalchemy==2.3.2
flask_login==0.4.1
bcrypt==3.1.6
requests==2.21.0
cryptography==2.8
pycryptodome==3.9.7